fix: 뒤바뀌어 있던 ui 파일 이름 올바르게 수정
line-manager-ui.js <-> section-manager-ui.js
파일이름과 클래스이름이 뒤바뀌어 있던 현상 수정
